Fixe problems with pasting to desktop Excel, changed e2e tests according to the change, moved some utils to another files
