+++ handsontable/src/plugins/collapsibleColumns/__tests__/collapsibleColumns.spec.js
@@ -2833,6 +2833,7 @@ describe('CollapsibleColumns', () => {
         '<meta name="generator" content="Handsontable"/>',
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
             '<tr>',
               '<th>A</th>',
@@ -2856,6 +2857,7 @@ describe('CollapsibleColumns', () => {
               '<td>J1</td>',
             '</tr>',
           '</tbody>',
+        '<!--EndFragment-->',
         '</table>'
       ].join(''));
       /* eslint-enable */
