+++ handsontable/src/plugins/copyPaste/__tests__/hooks/afterCopy.spec.js
@@ -39,6 +39,7 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>' +
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>' +
         '<table>' +
+        '<!--StartFragment-->' +
           '<tbody>' +
             '<tr>' +
               '<td>C2</td>' +
@@ -56,6 +57,7 @@ describe('CopyPaste', () => {
               '<td>E4</td>' +
             '</tr>' +
           '</tbody>' +
+          '<!--EndFragment-->' +
         '</table>',
       );
       /* eslint-enable */
@@ -88,6 +90,7 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>' +
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>' +
         '<table>' +
+        '<!--StartFragment-->' +
           '<thead>' +
             '<tr>' +
               '<th>C</th>' +
@@ -112,6 +115,7 @@ describe('CopyPaste', () => {
               '<td>E4</td>' +
             '</tr>' +
           '</tbody>' +
+        '<!--EndFragment-->' +
         '</table>',
       );
       /* eslint-enable */
@@ -153,6 +157,7 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>' +
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>' +
         '<table>' +
+        '<!--StartFragment-->' +
           '<thead>' +
             '<tr>' +
               '<th>C</th>' +
@@ -177,6 +182,7 @@ describe('CopyPaste', () => {
               '<td>E4</td>' +
             '</tr>' +
           '</tbody>' +
+        '<!--EndFragment-->' +
         '</table>',
       );
       expect(afterCopySpy.calls.argsFor(0)[0].getMetaInfo()).toEqual({
@@ -219,6 +225,7 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>' +
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>' +
         '<table>' +
+        '<!--StartFragment-->' +
           '<thead>' +
             '<tr>' +
               '<th>C</th>' +
@@ -247,6 +254,7 @@ describe('CopyPaste', () => {
               '<td>D4</td>' +
             '</tr>' +
           '</tbody>' +
+        '<!--EndFragment-->' +
         '</table>',
       );
