+++ handsontable/src/plugins/copyPaste/__tests__/methods/copyColumnHeadersOnly.spec.js
@@ -100,9 +100,16 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>',
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
-            '<tr><th>A-0-1</th><th>B-1-1</th><th>C-2-1</th><th>D-3-1</th></tr>',
+            '<tr>',
+              '<th>A-0-1</th>',
+              '<th>B-1-1</th>',
+              '<th>C-2-1</th>',
+              '<th>D-3-1</th>',
+            '</tr>',
           '</thead>',
+        '<!--EndFragment-->',
         '</table>',
       ].join(''));
       /* eslint-enable */
@@ -134,9 +141,15 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>',
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
-            '<tr><th>B</th><th>C</th><th>D</th></tr>',
+            '<tr>',
+              '<th>B</th>',
+              '<th>C</th>',
+              '<th>D</th>',
+            '</tr>',
           '</thead>',
+        '<!--EndFragment-->',
         '</table>',
       ].join(''));
       /* eslint-enable */
@@ -196,9 +209,15 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>' +
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
-            '<tr><th>B</th><th>C</th><th>D</th></tr>',
+            '<tr>',
+              '<th>B</th>',
+              '<th>C</th>',
+              '<th>D</th>',
+            '</tr>',
           '</thead>',
+        '<!--EndFragment-->',
         '</table>',
       ].join(''));
       /* eslint-enable */
@@ -258,9 +277,17 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>',
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
-            '<tr><th>A</th><th>B</th><th>C</th><th>D</th><th>E</th></tr>',
+            '<tr>',
+              '<th>A</th>',
+              '<th>B</th>',
+              '<th>C</th>',
+              '<th>D</th>',
+              '<th>E</th>',
+            '</tr>',
           '</thead>',
+        '<!--EndFragment-->',
         '</table>',
       ].join(''));
       /* eslint-enable */
