+++ handsontable/src/plugins/copyPaste/__tests__/methods/copyWithColumnHeaders.spec.js
@@ -57,12 +57,14 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>',
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
             '<tr><th>A-0-2</th></tr>',
           '</thead>',
           '<tbody>',
             '<tr><td>A2</td></tr>',
           '</tbody>',
+        '<!--EndFragment-->',
         '</table>',
       ].join(''));
       /* eslint-enable */
@@ -113,6 +115,7 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>' +
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
             '<tr><th>A-0-2</th><th>B-1-2</th><th>C-2-2</th></tr>',
           '</thead>',
@@ -120,6 +123,7 @@ describe('CopyPaste', () => {
             '<tr><td>A1</td><td>B1</td><td>C1</td></tr>',
             '<tr><td>A2</td><td>B2</td><td>C2</td></tr>',
           '</tbody>',
+        '<!--EndFragment-->',
         '</table>',
         /* eslint-enable */
       ].join(''));
@@ -170,6 +174,7 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>',
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
             '<tr><th>B-1-1</th><th>C-2-1</th><th>D-3-1</th></tr>',
           '</thead>',
@@ -177,6 +182,7 @@ describe('CopyPaste', () => {
             '<tr><td>B2</td><td>C2</td><td>D2</td></tr>',
             '<tr><td>B3</td><td>C3</td><td>D3</td></tr>',
           '</tbody>',
+        '<!--EndFragment-->',
         '</table>',
       ].join(''));
       /* eslint-enable */
@@ -202,14 +208,20 @@ describe('CopyPaste', () => {
       plugin.onCopy(copyEvent); // emulate native "copy" event
 
       expect(copyEvent.clipboardData.getData('text/plain')).toBe('B2\tC2\tD2\nB3\tC3\tD3');
+      /* eslint-disable indent */
       expect(copyEvent.clipboardData.getData('text/html')).toBe([
         '<meta name="generator" content="Handsontable"/>' +
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
-        '<table><tbody>',
-        '<tr><td>B2</td><td>C2</td><td>D2</td></tr>',
-        '<tr><td>B3</td><td>C3</td><td>D3</td></tr>',
-        '</tbody></table>',
+        '<table>',
+        '<!--StartFragment-->',
+          '<tbody>',
+            '<tr><td>B2</td><td>C2</td><td>D2</td></tr>',
+            '<tr><td>B3</td><td>C3</td><td>D3</td></tr>',
+          '</tbody>',
+        '<!--EndFragment-->',
+        '</table>',
       ].join(''));
+      /* eslint-enable */
     });
 
     it('should copy cells with the most-bottom column headers to the clipboard when all columns are hidden', () => {
@@ -258,6 +270,7 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>',
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
             '<tr><th>B-1-1</th><th>C-2-1</th><th>D-3-1</th></tr>',
           '</thead>',
@@ -265,6 +278,7 @@ describe('CopyPaste', () => {
             '<tr><td>B2</td><td>C2</td><td>D2</td></tr>',
             '<tr><td>B3</td><td>C3</td><td>D3</td></tr>',
           '</tbody>',
+        '<!--EndFragment-->',
         '</table>',
       ].join(''));
       /* eslint-enable */
@@ -291,14 +305,28 @@ describe('CopyPaste', () => {
       plugin.onCopy(copyEvent); // emulate native "copy" event
 
       expect(copyEvent.clipboardData.getData('text/plain')).toBe('B2\tC2\tD2\nB3\tC3\tD3');
+      /* eslint-disable indent */
       expect(copyEvent.clipboardData.getData('text/html')).toBe([
         '<meta name="generator" content="Handsontable"/>' +
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
-        '<table><tbody>',
-        '<tr><td>B2</td><td>C2</td><td>D2</td></tr>',
-        '<tr><td>B3</td><td>C3</td><td>D3</td></tr>',
-        '</tbody></table>',
+        '<table>',
+        '<!--StartFragment-->',
+        '<tbody>',
+          '<tr>',
+            '<td>B2</td>',
+            '<td>C2</td>',
+            '<td>D2</td>',
+          '</tr>',
+          '<tr>',
+            '<td>B3</td>',
+            '<td>C3</td>',
+            '<td>D3</td>',
+          '</tr>',
+        '</tbody>',
+        '<!--EndFragment-->',
+        '</table>',
       ].join(''));
+      /* eslint-enable */
     });
 
     it('should copy most-bottom column headers only to the clipboard when all rows are trimmed', () => {
@@ -345,9 +373,11 @@ describe('CopyPaste', () => {
         '<meta name="generator" content="Handsontable"/>',
         '<style type="text/css">td{white-space:normal}br{mso-data-placement:same-cell}</style>',
         '<table>',
+        '<!--StartFragment-->',
           '<thead>',
             '<tr><th>A-0-1</th><th>B-1-1</th><th>C-2-1</th><th>D-3-1</th><th>E-4-1</th></tr>',
           '</thead>',
+        '<!--EndFragment-->',
         '</table>',
       ].join(''));
       /* eslint-enable */
