+++ handsontable/src/selection/index.js
@@ -3,6 +3,7 @@ import { handleMouseEvent } from './mouseEventHandler';
 import {
   detectSelectionType,
   normalizeSelectionFactory,
+  normalizeRanges,
 } from './utils';
 
 export * from './highlight/highlight';
@@ -10,5 +11,6 @@ export {
   handleMouseEvent,
   Selection,
   detectSelectionType,
-  normalizeSelectionFactory
+  normalizeSelectionFactory,
+  normalizeRanges,
 };
